requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
python-dotenv==1.0.0
email-validator==2.1.0
geopy==2.4.0
requests==2.31.0
python-dateutil==2.8.2
Flask-Bootstrap==3.3.7.1
Flask-Moment==1.0.1
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
PyJWT>=2.0.0
